refactor: ProblemHeader 생성 및 분리
